Switch from Create to Link when conflict
I have made this change several years ago. Daniel reject this change
because user may lost a part of data. In fact, a merge will be better
but it's good for me like this...

When there is a conflict for person creation, a list of all homonyms is
proposed. User can witch to link to one of them instead of create.