Add place autocompletion
Added a new request DATALIST to get datalist contents from GeneWeb (not
limited to places). Cache max-age is configurable for it.
Added a datalist in updind and updfam for places.
Added asynchronous HTML include to load datalist in background.

TODO: A better implementation is needed to get a limited size for
datalist. It's needed to create a places table inside GeneWeb database
in order to search on it without to parse all persons and families.